refactor: extract IOC/enrichment/KEV/NSRL routes into routes/threatIntel.ts
Pure structural move out of createApp; no handler logic changed. Moved the 34
threat-intel routes (IOC add/bulk-enrich/bulk-tag/sources/provenance/
provenance-chain, global IOC whitelist + per-case IOC exclude, enrich +
enrich-control + enrich-health, customer-exposure, KEV, NSRL) into
routes/threatIntel.ts.

The enrichment engine and its background reachability poller stay in createApp;
the moved routes reach them through newly graduated RouteContext members:
enrichInBackground / autoEnrichIfEnabled / enabledProvidersFor (stable methods),
enrichPending() (live accessor for the poller-drained Set), and nsrlDb() /
setNsrlDb() (live accessor + mutator for the runtime-swappable NSRL RDS handle).
Reused existing ctx members (enrichHealth, applyWhitelistToCase, applyNsrlToCase,
resynthesizeInBackground). Domain-local stores rebuilt in-module.
